+++ classes/form/competency_dynamic_form.php
@@ -119,7 +119,10 @@ class competency_dynamic_form extends \core_form\dynamic_form {
      * @return \moodle_url
      */
     protected function get_page_url_for_dynamic_submission(): \moodle_url {
-        return new \moodle_url('/local/dimensions/central.php', ['frameworkid' => $this->get_frameworkid()]);
+        return \local_dimensions\helper::hub_page_url(
+            $this->get_context_for_dynamic_submission(),
+            ['frameworkid' => $this->get_frameworkid()]
+        );
     }
 
     /**