+++ classes/form/template_dynamic_form.php
@@ -99,7 +99,7 @@ class template_dynamic_form extends \core_form\dynamic_form {
      * @return \moodle_url
      */
     protected function get_page_url_for_dynamic_submission(): \moodle_url {
-        return new \moodle_url('/local/dimensions/central.php');
+        return \local_dimensions\helper::hub_page_url($this->get_context_for_dynamic_submission());
     }
 
     /**